+++ babel.config.js
@@ -1,23 +1,34 @@
 const { babelAliases } = require('./config/aliases');
 
-module.exports = {
-    presets: [
-        [
-            '@babel/env',
-            {
-                modules: process.env.BABEL_ENV === 'es' ? false : 'auto',
-            },
-        ],
-    ],
+const config = {
+    presets: [],
     plugins: [
-        '@babel/plugin-proposal-object-rest-spread',
         [
             'babel-plugin-module-resolver',
             {
                 alias: babelAliases,
             },
         ],
-        '@babel/plugin-transform-runtime',
     ],
     ignore: ['**/__tests__/', '**/*.test.js'],
 };
+
+if (process.env.BABEL_ENV === 'es') {
+    config.presets.push([
+        '@babel/modules',
+        {
+            loose: true,
+        },
+    ]);
+} else {
+    config.presets.push([
+        '@babel/env',
+        {
+            loose: true,
+        },
+    ]);
+
+    config.plugins.push(...['@babel/plugin-proposal-object-rest-spread', '@babel/plugin-transform-runtime']);
+}
+
+module.exports = config;