+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 2.1.1 - 2019-11-15
+
+### Updated
+
+-   Upgrade dependencies.
+-   Babel config
+    -   Use `@babel/preset-modules` for `/es` output -> decreased package size.
+    -   Enable `loose` option -> decreased package size.
+
 ## 2.1.0 - 2019-11-11
 
 ### Updated